"check_style_score -> 3 of 4 passed. style score: 0/1 "

+++ CSPP2-assignments/m2/Assignment-2/Assignment-2/Solution.java
@@ -13,7 +13,7 @@ public class Solution {
 		//System.out.println(a + " " + b);
 	}
 	public static void rootsOfQuadraticEquation(int a, int b, int c) {
-		int determinant = (b*b) - (4*a*c);
+		double determinant = (b*b) - (4*a*c);
 		if (determinant > 0){
 			double root1 = (-b + Math.sqrt(determinant))/(2*a);
 			double root2 = (-b - Math.sqrt(determinant))/(2*a);
